Added Main Menu and altered game structure
Added a Main Menu to start the game.

Made the dungeon portion of the game into a scene, allowing for it to be created upon start of the game. This will save complications massively as it doesn't have to sit in the background during main menu. Will be useful down the line.

Added a new font, m5x7 by Daniel Linssen.